split parser from new functionality
